Move LayertreeController to a separate file
In order to have one goog.provide per file.
